Change samples to ints instead of longs
We will never support more than 2^31 samples per node and avoid casting to int at the end

+++ algo/src/main/java/org/neo4j/gds/embeddings/graphsage/Layer.java
@@ -27,7 +27,7 @@ import java.util.List;
 import java.util.stream.LongStream;
 
 public interface Layer {
-    long sampleSize();
+    int sampleSize();
 
     Aggregator aggregator();
